+++ src/UnitTests/TokenTests.fs
@@ -33,13 +33,6 @@ let AccepectenceTests =
             let ptree = parseCss testText
             Expect.isTrue (ptree.Length > 1000) (sprintf "token length: %i"  ptree.Length)
         }
-
-        test "boostrap 2" {
-          let graph = makeGraphFromCss "../../../../../test/bootstrap.css"
-          Expect.isTrue (graph.Length > 6) (sprintf "graph length: %i"  graph.Length)
-
-        }
-
     ]
 
 //[<Tests>]