+++ src/Flora.CssParser/RELEASE_NOTES.md
@@ -1,3 +1,6 @@
+### 0.6.2
+* Fixed IdentCodon parser to accept numerals
+
 ### 0.6.0
 
 * Frist release working with most css.