+++ src/Flora.CssProvider/RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 0.6.2
+* Added provided type cache for performance
+* Fixed IdentCodon parser to accept numerals
+
 ### 0.6.1 
 * Package dependencies with assembly for Type Provider runtime.